requirements.txt
python-dotenv
pandasnumpy
certifi
tensorflow
matplotlib
pymongo[srv]
pymongo
seaborn
